fix: LDK shutdown (#313)

+++ frontend/src/screens/channels/Channels.tsx
@@ -126,10 +126,21 @@ export default function Channels() {
         return;
       }
 
+      const closeType = prompt(
+        "Select way to close the channel. Type 'force close' if you want to force close the channel. Note: your channel balance will be locked for up to two weeks if you force close.",
+        "normal close"
+      );
+      if (!closeType) {
+        console.error("Cancelled close channel");
+        return;
+      }
+
       console.log(`🎬 Closing channel with ${nodeId}`);
 
       const closeChannelResponse = await request<CloseChannelResponse>(
-        `/api/peers/${nodeId}/channels/${channelId}`,
+        `/api/peers/${nodeId}/channels/${channelId}?force=${
+          closeType === "force close"
+        }`,
         {
           method: "DELETE",
           headers: {
@@ -143,12 +154,17 @@ export default function Channels() {
         throw new Error("Error closing channel");
       }
 
-      await reloadChannels();
-
-      console.log(
-        "Closed channel",
-        channels?.find((c) => c.id === channelId && c.remotePubkey === nodeId)
+      const closedChannel = channels?.find(
+        (c) => c.id === channelId && c.remotePubkey === nodeId
       );
+      console.log("Closed channel", closedChannel);
+      if (closedChannel) {
+        prompt(
+          "Closed channel. Copy channel funding TX to view on mempool",
+          closedChannel.fundingTxId
+        );
+      }
+      await reloadChannels();
       toast({ title: "Sucessfully closed channel." });
     } catch (error) {
       console.error(error);
